refactored utils
- seperated first_index and replace_words into util
- aggregated the relevant utils into string_utils so there wont be a million files with 3 lines of code each
